requirements.txt
Flask==2.3.3
pandas==1.5.3
openpyxl==3.1.2
fpdf2==2.7.4
gunicorn==21.2.0